server: Try to handle the TAKClient less, delegate to router
The TAKClient is rather thread-unsafe right now. When the server thread
feeds it data, it can change the TAKClient while the router is trying to
process information.

There is another issue with the TAKClient socket being closed in the
server, and some potential issues with the COTRouter stumbling because
of that... TBD.