models: Improve error handling with from_evt
There are a number of nasty errors that can happen when converting XML
to a Python object. Instead of trying to catch ValueError, TypeError,
and all sorts of other unknowns, we will reduce it to UnmarshalError.

Additionally, in the event that the client has an unhandled parsing
error, we log that as well (and most importantly, do not crash!)