takuser: Send the detail as received from the user
This was a big effort. The idea here is that we will not understand
everything that the TAK client sends us, like custom extensions. We
should not (and cannot) presume to have a full understanding of what
fields will be present.

If we generate a TAKUser detail, there are some fields we know should
be present. But if we are simply routing one, send all the fields, even
if we do not understand them.