Add a basic docker setup for development
This doesn't keep a copy of the code in the container - it's mounted in at runtime to make testing and iteration quicker.
takyctl setup is run automatically on first run to create certificates and config, then the /etc/taky directory is mounted out to the local ./docker directory so the config can be edited.

Usage:

```
docker-compose build
docker-compose up
```