Data Package Privacy - respect the `tool` setting that ATAK sets on upload
If you send a package directly to the server, it will be set to "public"
and will show in the listing of available files.
If you send a packages to another user via the server it will be set to
"private" and not show in the listing.

Note that privacy is not enforced on the download - if you know the URL
you can get the package, regardless of the visibility setting.