cot.router: Send packet to all sockets matching CS/UID
This is a bit of a weird change. Right now, taky doesn't enforce unique
UIDs and Callsigns for clients. If two clients connect and claim to be
JENNY, they should both get the packet. This may be problematic,
especially if both clients claim to have the same UID.

For now, given that taky doesn't try to secure the COT network, we will
not kick off clients with duplicate callsigns or UIDs.