Unify code and configuration to use log_cot
Originally, log_cot was a boolean, but then it became a variable that
pointed to a directory. Unfortunately, the code frequently used
"cot_log", and I made a mistake in the documentation. (Additionally, in
a test case!)

To preserve user configurations and setups, the all references to
strings containing "cot_log" were changed to "log_cot" in the code base
as well.

Closes issue #37