config: Set the config path when we read the file
Fixes #79. cfg_path was a terrible decision, and really shouldn't exist.
The idea was for the config object to keep track of where it was loaded
from -- but I'm abusing the ever loving stuffing out of it, and it
should be removed in a future release.

For the time being, however, I'm moving this line within the conditional
to ensure that cfg_path is only set when we are directly reading a file
on the disk. It's pedantic, but hopefully this will prevent any further
undefined behavior until I can revamp the config system.